+++ tests/testthat/test-test-path.R
@@ -1,11 +1,10 @@
 context("test_path")
 
-test_that("throws error if file doesn't exist", {
-  expect_error(test_path("XXXX"), "doesn't exist")
-})
-
 test_that("returns local path when called in tests", {
   expect_equal(test_path("test-test-path.R"), "test-test-path.R")
+
+  # even if path doesn't (yet) exists
+  expect_equal(test_path("xxxx"), "xxxx")
 })
 
 test_that("returns full path when called outside tests", {
@@ -13,6 +12,8 @@ test_that("returns full path when called outside tests", {
   on.exit(setwd(old))
 
   expect_equal(test_path("empty"), "tests/testthat/empty")
+  # even when file doesn't exist
+  expect_equal(test_path("xxx"), "tests/testthat/xxx")
 })
 
 test_that("throws error if can't find tests/testthat", {
@@ -20,5 +21,4 @@ test_that("throws error if can't find tests/testthat", {
   on.exit(setwd(old))
 
   expect_error(test_path("empty"), "Can't find `tests/testthat/`")
-
 })